docs: tighten invite-links recipe after review
- Align naming with sibling recipes: snake_case `user_id`, `$createdBy`
  instead of a hand-rolled `createdBy` column, drop unused `isPublic`.
- Split secondary "session claim" pattern into its own snippet files so
  each file has a single `definePermissions` block on a single app.
- Reframe the secondary pattern as a session claim (which is what the
  client actually attaches), drop the misleading "mint a JWT" wording
  and the no-op `authMode: "external"`.
- Add a Callout explaining why payload claims can't escalate privilege:
  server-verified identity and pre-existing claims take precedence over
  payload values during the merge in sync_manager.
